Revamp build process to separate locale data
This changes the build process of this package to separate the locale data from the runtime lib.

When using this package in Node.js, the library will still be primed with all its locale data. _But_ on the client, it won't. The `dist/` dir now has the following structure:

```
dist/
├── intl-messageformat-with-locales.js
├── intl-messageformat-with-locales.min.js
├── intl-messageformat.js
├── intl-messageformat.js.map
├── intl-messageformat.min.js
└── locale-data
    ...
    ├── en.js
    ...
    ├── fr.js
    ...
    └── yo.js

1 directory, 157 files
```